requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.2.0
Pillow==10.0.0
scikit_learn==1.1.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
